Week2/Day5/Hackathon/Car Dealership Management/requirements.txt

Flask==2.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
psycopg2-binary==2.8.6
WTForms==3.0.1
# Pillow is optional - install with: pip install Pillow or conda install pillow
# Pillow==9.5.0
python-dotenv==0.21.0
